numbers_prog/requirements.txt
psycopg2-binary==2.9.1
google-api-python-client==2.51.0
oauth2client==4.1.3
httplib2==0.20.4
beautifulsoup4==4.9.1
requests==2.26.0

